fix(app): render the pending-profile gate inside WorkspaceProvider
Re-applies #985 (reverted in #986). The waiting screen was a hook called
from MainApp, which renders above WorkspaceProvider — useWorkspace threw
and every IDE page showed "Mako hit a display error". It is now
PendingProfileGate, a component mounted in AuthWrapper under the
provider, wrapping the protected tree.
