+++ app/src/components/AuthWrapper.tsx
@@ -2,6 +2,7 @@ import { AuthProvider } from "../contexts/auth-context";
 import { WorkspaceProvider } from "../contexts/workspace-context";
 import { OnboardingProvider } from "../contexts/onboarding-context";
 import { ProtectedRoute } from "./ProtectedRoute";
+import { PendingProfileGate } from "./PendingProfileGate";
 
 interface AuthWrapperProps {
   children: React.ReactNode;
@@ -19,7 +20,9 @@ export function AuthWrapper({ children }: AuthWrapperProps) {
     <AuthProvider>
       <OnboardingProvider>
         <WorkspaceProvider>
-          <ProtectedRoute>{children}</ProtectedRoute>
+          <ProtectedRoute>
+            <PendingProfileGate>{children}</PendingProfileGate>
+          </ProtectedRoute>
         </WorkspaceProvider>
       </OnboardingProvider>
     </AuthProvider>
